// src/app/interfaces/mazars.model.ts
export interface IUser {
  id: string;
  imagePath: any;
  firstName: string;
  lastName: string;
  company: string;
  email: string;
  password: string;
  confirmPassword: string;
  condition1: boolean;
  condition2: boolean;
  source: number;
  userCategoryName: string;
  userGroupName: string;
  preferredLanguage: string;
}